+++ apps/block_scout_web/lib/block_scout_web/controllers/tokens/bridged_tokens_controller.ex
@@ -28,6 +28,16 @@ defmodule BlockScoutWeb.BridgedTokensController do
           )
       end
 
+    items_count_str = Map.get(params, "items_count")
+
+    items_count =
+      if items_count_str do
+        {items_count, _} = Integer.parse(items_count_str)
+        items_count
+      else
+        0
+      end
+
     items =
       tokens_page
       |> Enum.with_index(1)
@@ -37,7 +47,7 @@ defmodule BlockScoutWeb.BridgedTokensController do
           "_tile.html",
           token: token,
           bridged_token: bridged_token,
-          index: index
+          index: items_count + index
         )
       end)